:lipstick: add redirection
