+++ src/styles/DropDownButton.ts
@@ -1,0 +1,22 @@
+import { StyleSheet } from "react-native";
+
+const styles = StyleSheet.create({
+	container: {
+		position: "absolute",
+		top: 20,
+		left: 40,
+		zIndex: 2
+	},
+	menu: {
+		position: "absolute",
+		backgroundColor: "#313131",
+		top: 10,
+		left: -110,
+		zIndex: 5
+	},
+	logoutText: {
+		color: "#FF6B6B"
+	}
+});
+
+export default styles;
